Spinny Bones!
The bones spin when you throw them! I am so happy! It is so good to see a bone spinning through the air before it clochks on an animals head. The caveman in me is happy. It was a little weird to get the bone spawning in the right place, but it works. So cool.