Trying to use events!
I'm trying to get the indicators to find the ball when I spawn a new one. It works perfectly for the enemy, but the indicators don't respond at all, and I don't know why. They did once, but then I must have changed something, because they don't anymore. THE CODE IS LITERALLY THE SAME AS IN THE ENEMY, THEY'RE BEING STUPID.