Object Pooling
Kinda broke the game. The bone spawns several feet in front of the player, and because I wanted to make the bone spin, I had multiple prefabs that got kinda weird, and they didn't work well with the object pooling script. I think I want to go back to the backup I made.